调整为1积分=0.002元

+++ src/main/java/com/jiazhe/youxiang/server/vo/resp/project/ProjectResp.java
@@ -8,6 +8,8 @@ package com.jiazhe.youxiang.server.vo.resp.project;
 import com.jiazhe.youxiang.server.vo.BaseVO;
 import io.swagger.annotations.ApiModelProperty;
 
+import java.math.BigDecimal;
+
 /**
  * 在这里编写类的功能描述
  *
@@ -26,7 +28,7 @@ public class ProjectResp extends BaseVO {
     @ApiModelProperty("排序")
     private Integer priority;
     @ApiModelProperty("积分兑换比例")
-    private Integer pointConversionRate;
+    private BigDecimal pointConversionRate;
     @ApiModelProperty("项目状态")
     private Integer status;
 
@@ -62,11 +64,11 @@ public class ProjectResp extends BaseVO {
         this.priority = priority;
     }
 
-    public Integer getPointConversionRate() {
+    public BigDecimal getPointConversionRate() {
         return pointConversionRate;
     }
 
-    public void setPointConversionRate(Integer pointConversionRate) {
+    public void setPointConversionRate(BigDecimal pointConversionRate) {
         this.pointConversionRate = pointConversionRate;
     }
